refactor: Remove deprecated environment service interface
