+++ src/modules/enviroment/application/interface/enviroment.base.interface.ts
@@ -1,0 +1,11 @@
+export interface IEnviromentValues {
+  name: string;
+  value: string;
+  collectionId: string;
+}
+
+export interface IUpdateEnviromentValues extends Partial<IEnviromentValues> {
+  id: string;
+}
+
+export const ENVIROMENT_SERVICE = 'ENVIROMENT_SERVICE';
